Fixes #34 logged on Knox-mpu
The "path" in a HTTP request uses space as a delimiter. If a file name contains a space, the HTTP request will be,
    GET /New Text Document.txt HTTP/1.1
This will result in AWS returning an "Unsupported HTTP version" error.
The fix is to replace the spaces in the "path" with the url encoded string "%20". The new request will now be,
    GET /New%20Text%20Document.txt HTTP/1.1
Note: We cannot use the native JS encodeURI(). This is because the file name can have a '/' to indicate the folder structure of your S3 bucket.
For Ex. If you are uploading a picture, to a folder called "pictures" in your S3 bucket, you will have to specify the file name as "/pictures/New Pic.jpg".
The encoded version of this will be "%2Fpictures%2FNew%20Pic.jpg".
If you take a look at "Client.prototype.request" in client.js in the Knox library, you will see that if a file name does not start with a '/', Knox adds one.
This will result in a file named "%2Fpictures%2FNew%20Pic.jpg" being uploaded to the root directory of your S3 bucket.